+++ app/sitemap.ts
@@ -9,7 +9,7 @@ export default function sitemap(): MetadataRoute.Sitemap {
   return [
     {
       url: BASE_URL,
-      lastModified: new Date('2026-07-30'),
+      lastModified: new Date('2026-08-22'),
       changeFrequency: 'weekly',
       priority: 1.0,
     },